Split the configuration file management in multiple modules (#1996)
Fix #1790
